add initial version of ebclfsa qemu image and demo application
This adds an image with two VMs running on the hypervisor.
The low integrity vm is a usual baseos vm with crinit and elos.
The high integrity vm has no applications installed at the moment. Just the
minimal bootstrap rootfs with busybox is installed and the kernel falls back to starting sh.

Co-authored-by: Anton Jahn <anton.jahn@elektrobit.com>