taskfiles: Ensure qemu reboots without force building
When one builds with task a qemu image and then poweroff the qemu and
wishes to restart it executing task will just assume that all is
uptodate as it was just done. While we do not want to rebuild the image,
we of course want to restart qemu so use method: none to ensure the
tasks are re-run, all other build task will be skiped as their generates
are unchanged in this situation.

Signed-off-by: Simone Weiß <simone.p.weiss@posteo.com>