cmake buildsystem PoC
This showcases the build of three images (appdev for x86 and arm and ebclfsa) using cmake.

Common functions used are in cmake/ebcl_sdk.cmake. This file should end up in the devcontainer,
if the PoC leaves its PoC-state.

The basic design idea is to follow the way cmake works for C/C++ builds, i.e.:
Everything that creates an artifact is a target (a rootfs, an initrd, ...).
These targets are then used in other targets (for example in the image/embdgen build).

All paths for intermediate files are handled internally (like root.tar, root.config.tar).
The path to the files (or directories) are stored in properties of the target, so they can be
fetched by a later build step.
This leads to the problem, that these paths are referenced in other yaml-config files.
For this PoC it is solved by adding a preprocessing step (cmake's configure_file) to insert
the paths into the yaml files. All dependencies are directly visible in the CMakeLists.txt (See ARTIFACT and OUTPUT_DIR).
For a finished solution I'd suggest adding a variable-feature to the build tools, that allows defining
variables on the command line, that substitute placeholders in the yaml files.

To build any of the variants:
 * Create a directory in an arbitrary location and cd into it
 * run cmake <source_dir>
 * run make (or make qemu)
That's it, all logfiles can be found in the buildiir usually in a dir called <target>.target